# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/v2/contract/accounts/{address}": {
            "get": {
                "description": "查询合约账户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-contract"
                ],
                "summary": "查询合约账户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "合约地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.V2ContractActResult"
                        }
                    }
                }
            }
        },
...
        "/v2/contract/events/{txhash}": {
            "get": {
                "description": "查询事件日志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-contract"
                ],
                "summary": "查询事件日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "交易hash",
                        "name": "txhash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.V2EvmLog"
                        }
                    }
                }
            }
        },
...
        "/v3/dpos/poollogs": {
            "get": {
                "description": "查询矿池记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-dpos"
                ],
                "summary": "查询矿池记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "高度",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.DPOSPoolLog"
                            }
                        }
                    }
                }
            }
        },
        "/v3/dpos/ranklogs": {
            "get": {
                "description": "查询DPOS排名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-dpos"
                ],
                "summary": "查询DPOS排名",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地址",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "高度",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.DPOSRankLog"
                            }
                        }
                    }
                }
            }
        },
        "/v3/dpos/tcnlogs": {
            "get": {
                "description": "查询节点dpos记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-dpos"
                ],
                "summary": "查询节点dpos记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地址",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "高度",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.DPOSTcnLog"
                            }
                        }
                    }
                }
            }
        },
        "/v3/dpos/tinlogs": {
            "get": {
                "description": "查询用户dpos记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-dpos"
                ],
                "summary": "查询用户dpos记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地址",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "高度",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.DPOSTinLog"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bean.V2ContractActResult": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "balance": {
                    "description": "余额",
                    "type": "string"
                },
                "code": {
                    "description": "合约字节码",
                    "type": "string"
                },
                "nonce": {
                    "description": "nonce",
                    "type": "integer"
                },
                "suicided": {
                    "description": "合约是否已自杀",
                    "type": "boolean"
                }
            }
        },
...
        "bean.V2EvmLog": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "合约地址",
                    "type": "string"
                },
                "blockHash": {
                    "description": "区块hash",
                    "type": "string"
                },
                "blockNumber": {
                    "description": "区块高度",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "logIndex": {
                    "description": "日志索引",
                    "type": "integer"
                },
                "removed": {
                    "description": "是否已移除",
                    "type": "boolean"
                },
                "topics": {
                    "description": "事件主题",
                    "type": "string"
                },
                "transactionHash": {
                    "description": "交易hash",
                    "type": "string"
                },
                "transactionIndex": {
                    "description": "交易索引",
                    "type": "integer"
                }
            }
        },
...
        "types.DPOSPoolLog": {
            "type": "object",
            "properties": {
                "balance": {
                    "description": "矿池余额",
                    "type": "string"
                },
                "blockAt": {
                    "description": "区块时间",
                    "type": "string"
                },
                "height": {
                    "description": "去看高度",
                    "type": "integer"
                },
                "id": {
                    "description": "数据库自增id",
                    "type": "integer"
                },
                "mined": {
                    "description": "挖矿量",
                    "type": "string"
                },
                "released": {
                    "description": "实际释放量，实际释放量约等于挖矿量",
                    "type": "string"
                },
                "total": {
                    "description": "总股权",
                    "type": "string"
                }
            }
        },
        "types.DPOSRankLog": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "节点地址",
                    "type": "string"
                },
                "blockAt": {
                    "description": "区块时间",
                    "type": "string"
                },
                "height": {
                    "description": "区块高度",
                    "type": "integer"
                },
                "id": {
                    "description": "数据库自增id",
                    "type": "integer"
                },
                "mortgaged": {
                    "description": "节点抵押量",
                    "type": "string"
                },
                "rank": {
                    "description": "排名",
                    "type": "integer"
                },
                "total": {
                    "description": "总股权",
                    "type": "string"
                },
                "voted": {
                    "description": "用户投票量",
                    "type": "string"
                },
                "voters": {
                    "description": "投票用户数",
                    "type": "integer"
                }
            }
        },
        "types.DPOSTcnLog": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "节点地址",
                    "type": "string"
                },
                "blockAt": {
                    "description": "区块时间",
                    "type": "string"
                },
                "height": {
                    "description": "区块高度",
                    "type": "integer"
                },
                "id": {
                    "description": "数据库自增id",
                    "type": "integer"
                },
                "mortgaged": {
                    "description": "节点抵押量",
                    "type": "string"
                },
                "profit": {
                    "description": "收益",
                    "type": "string"
                },
                "voted": {
                    "description": "用户投票量",
                    "type": "string"
                },
                "voters": {
                    "description": "投票用户数",
                    "type": "integer"
                }
            }
        },
        "types.DPOSTinLog": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "用户地址",
                    "type": "string"
                },
                "blockAt": {
                    "description": "区块时间",
                    "type": "string"
                },
                "height": {
                    "description": "区块高度",
                    "type": "integer"
                },
                "id": {
                    "description": "数据库自增id",
                    "type": "integer"
                },
                "mortgaged": {
                    "description": "用户抵押量",
                    "type": "string"
                },
                "profit": {
                    "description": "收益",
                    "type": "string"
                },
                "validator": {
                    "description": "用户选举的节点地址",
                    "type": "string"
                }
            }
        }
    }
}